recruitai-backend-master/requirements.txt
absl-py==1.0.0
amqp==5.1.0
appdirs==1.4.4
asgiref==3.5.0
astunparse==1.6.3
audioread==2.1.9
backports.zoneinfo==0.2.1
billiard==3.6.4.0
cachetools==5.0.0
celery==5.2.6
certifi==2021.10.8
cffi==1.15.0
chardet==3.0.4
charset-normalizer==2.0.12
click==8.1.2
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
decorator==4.4.2
Django
djangorestframework==3.13.1
docx2txt==0.8
fire==0.4.0
flatbuffers==2.0
fonttools==4.32.0
gast==0.5.3
gcloud==0.17.0
google-auth==2.6.3
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
googleapis-common-protos==1.56.0
grpcio==1.44.0
h5py==3.6.0
httplib2==0.20.4
idna==2.7
imageio==2.16.1
imageio-ffmpeg==0.4.5
importlib-metadata==4.11.3
joblib==1.1.0
jws==0.1.3
keras==2.8.0
Keras-Preprocessing==1.1.2
kombu==5.2.4
libclang==13.0.0
librosa==0.9.1
llvmlite==0.38.0
lxml==4.8.0
Markdown==3.3.6
moviepy
numba==0.55.1
numpy==1.21.0
oauth2client==3.0.0
oauthlib==3.2.0
opencv-python==4.5.5.64
opt-einsum==3.3.0
packaging==21.3
pdf2docx==0.5.3
Pillow==9.1.0
pooch
proglog==0.1.9
prompt-toolkit==3.0.29
protobuf==3.20.1rc1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pycryptodome==3.4.3
PyMuPDF==1.19.6
pyparsing==3.0.8
Pyrebase
python-docx==0.8.11
python-jwt==2.0.1
pytz==2022.1
requests
requests-oauthlib==1.3.1
requests-toolbelt==0.7.0
resampy==0.2.2
rsa==4.8
scikit-learn==1.0.2
scipy
six==1.16.0
SoundFile==0.10.3.post1
SpeechRecognition==3.8.1
sqlparse==0.4.2
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow-io-gcs-filesystem==0.24.0
termcolor==1.1.0
tf-estimator-nightly==2.8.0.dev2021122109
threadpoolctl==3.1.0
tqdm==4.64.0
typing-extensions==4.1.1
urllib3==1.23
vine==5.0.0
wcwidth==0.2.5
Werkzeug==2.1.1
wrapt==1.14.0
zipp==3.8.0
